+++ Nickel/Framework/ModStorageManager.cs
@@ -1,6 +1,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
+using Nickel.Common;
 using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;